refactor: split scan artifacts by what they do to an artifact set
lib/scan/artifacts.ts was 900 lines and 48 declarations spanning the whole
R2 lifecycle. Now a directory behind a barrel, so none of the 16 consumers
changed:

- types.ts    stored shapes and caps
- keys.ts     org-scoped key layout — an artifact read is scoped by
              construction, not by a check the caller must remember
- json-io.ts  the digest-verified put/get everything else goes through
- parse.ts    tolerant parsers for bodies read back, all null-returning so
              one unreadable artifact degrades to the D1 fallback
- write.ts / read.ts / delete.ts   the three operations

Verified with the declaration census rather than a green suite: all 48
names present in exactly one module, none dropped, none duplicated.
knip/build/verify clean. docs checked — no doc names this path.
